Fixed Issues with Collisions
Debugging: Using Print() to view cords of plays, player 2 y cords were in the decimals which caused a low chance of collision on the exact decimal number.
To fix this issue both players were moved to the left side of the window and both head east.
The issue was caused by the turtle's bad code where player 2 needed to be rotated at start-up, the rotation wasn't exact and was on an angle.