Updated Instructions and added Errors
Modified the README file, this will be the file the user will open for guidance on how to run python code. It gives the user insight on the minimum system requirements, how to install python and the game and some common questions. I've also modified the Main menu to check if the user is using an older version of Python. This is because the Tkinter module in 2.x is different to 3.x.